move windmill png to icons folder
