+++ pages/circulaire-windturbines/samenhang-maatregelen.js
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import Image from 'next/image';
 import Layout from '../../components/layouts/layout';
 import SamenhangLayout from '../../components/layouts/samenhang-layout';
-import WindmillIcon from '../../public/windmill.png';
+import WindmillIcon from '../../public/icons/windmill.png';
 import diagram from '../../public/windmill-samg.svg';
 
 export default function InfoPage() {
