+++ pages/circulaire-windturbines/welke-overheid-heeft.js
@@ -1,7 +1,7 @@
 import client from '../../lib/sanity';
 import Layout from '../../components/layouts/layout';
 import WelkeLayout from '../../components/layouts/welke-layout';
-import WindmillIcon from '../../public/windmill.png';
+import WindmillIcon from '../../public/icons/windmill.png';
 import windmillImageMob from '../../public/bevoegdheden/bevoegdheden-windmill.png';
 import { creatQuery } from '../../lib/queries';
