+++ pages/measures/windturbines.js
@@ -1,7 +1,7 @@
 import Layout from '/components/layouts/layout';
 import MeasuresLayout from '../../components/layouts/measures-layout';
 import client from '../../lib/sanity';
-import WindmillIcon from '../../public/windmill.png';
+import WindmillIcon from '../../public/icons/windmill.png';
 
 const lawsQuery = `
 *[_type == "measure" && thema == "circulaire-windturbines"]
